add part of infrastructure layer and first of api layer

+++ SimpleFinance.Application/Dto/StockDto.cs
@@ -5,6 +5,6 @@ public class StockDto
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Exchange { get; set; } = string.Empty;
-    public float ActualPrice { get; set; }
+    public double ActualPrice { get; set; }
     public IEnumerable<StockActionDto> Items { get; set; } = new LinkedList<StockActionDto>();
 }
